# perf: cache statements