# Command

```
RUST_LOG=error ./target/release/lam example serve --name store
```

## Before

```
Running 10s test @ http://localhost:3000
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   643.71us  143.38us   2.27ms   75.18%
    Req/Sec     7.75k   349.89     8.39k    69.31%
  155700 requests in 10.10s, 18.01MB read
Requests/sec:  15415.39
Transfer/sec:      1.78MB
```

## After

```
Running 10s test @ http://localhost:3000
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   511.43us  101.64us   7.33ms   81.78%
    Req/Sec     9.73k   375.50    10.62k    64.36%
  195488 requests in 10.10s, 22.64MB read
Requests/sec:  19356.43
Transfer/sec:      2.24MB
```